Create Abstract Evaluator that regroup Evaluator classic attributes and methods
