+++ evaluator/Rules/rules_evaluator.py
@@ -1,40 +1,16 @@
 from abc import *
 
-from config.cst import *
+from evaluator.abstract_evaluator import AbstractEvaluator
 
 
-# TODO : temp class
-class RulesEvaluator:
+class RulesEvaluator(AbstractEvaluator):
     __metaclass__ = ABCMeta
 
     def __init__(self):
         super().__init__()
-        self.logger = None
-        self.config = None
-        self.eval_note = START_EVAL_NOTE
-        self.pertinence = START_EVAL_PERTINENCE
-        self.logger = None
-        self.enabled = True
-
-    @classmethod
-    def get_name(cls):
-        return cls.__name__
-
-    def set_logger(self, logger):
-        self.logger = logger
-
-    def set_config(self, config):
-        self.config = config
-
-    def get_eval_note(self):
-        return self.eval_note
-
-    def get_pertinence(self):
-        return self.pertinence
-
-    def get_is_enabled(self):
-        return self.enabled
 
     @abstractmethod
-    def eval(self):
-        raise NotImplementedError("Eval not implemented")
+    def eval_impl(self) -> None:
+        raise NotImplementedError("Eval_impl not implemented")
+
+
