+++ evaluator/Social/social_evaluator.py
@@ -4,17 +4,16 @@ from abc import *
 from config.cst import *
 from botcore.config.config import load_config
 
+from evaluator.abstract_evaluator import AbstractEvaluator
 
-class SocialEvaluator(threading.Thread):
+
+class SocialEvaluator(AbstractEvaluator, threading.Thread):
     __metaclass__ = ABCMeta
 
     def __init__(self):
         threading.Thread.__init__(self)
         self.social_config = None
-        self.config = None
-        self.logger = None
-        self.symbol = None
-        self.history_time = None
+
         self.eval_note = START_EVAL_NOTE
         self.logger = None
         self.pertinence = START_EVAL_PERTINENCE
@@ -29,10 +28,6 @@ class SocialEvaluator(threading.Thread):
     @classmethod
     def get_config_file_name(cls):
         return SPECIFIC_CONFIG_PATH + cls.__name__ + CONFIG_FILE_EXT
-
-    @classmethod
-    def get_name(cls):
-        return cls.__name__
 
     def stop(self):
         self.keep_running = False
@@ -51,32 +46,8 @@ class SocialEvaluator(threading.Thread):
         else:
             self.set_default_config()
 
-    def set_logger(self, logger):
-        self.logger = logger
-
-    def set_symbol(self, symbol):
-        self.symbol = symbol
-
-    def set_history_time(self, history_time):
-        self.history_time = history_time
-
-    def set_config(self, config):
-        self.config = config
-
-    def get_eval_note(self):
-        return self.eval_note
-
-    def get_is_enabled(self):
-        return self.enabled
-
     def get_is_threaded(self):
         return self.is_threaded
-
-    def get_pertinence(self):
-        return self.pertinence
-
-    def get_is_updating(self):
-        return self.is_updating
 
     # to implement in subclasses if config necessary
     # required if is_threaded = False --> provide evaluator refreshing time
@@ -86,28 +57,8 @@ class SocialEvaluator(threading.Thread):
     def get_social_config(self):
         return self.social_config
 
-    # generic social eval that will call the indicator eval()
-    # and provide a safe execution by disabling multi-call
-    def eval(self):
-        self.is_updating = True
-        try:
-            self.eval_impl()
-        except Exception as e:
-            self.logger.error("Exception in eval_impl(): "+str(e))
-        finally:
-            self.is_updating = False
-
-    # eval new data
-    # Notify if new data is relevant
-    # example :
-    # def eval_impl(self):
-    #   for post in post_selected
-    #       note = sentiment_evaluator(post.text)
-    #       if(note > 10 || note < 0):
-    #           self.need_to_notify = True
-    #       self.eval_note += note
     @abstractmethod
-    def eval_impl(self):
+    def eval_impl(self) -> None:
         raise NotImplementedError("Eval_impl not implemented")
 
     # get data needed to perform the eval
