+++ evaluator/TA/TA_evaluator.py
@@ -1,37 +1,20 @@
 from abc import *
 
 from config.cst import *
+from evaluator.abstract_evaluator import AbstractEvaluator
 
 
-class TAEvaluator:
+class TAEvaluator(AbstractEvaluator):
     __metaclass__ = ABCMeta
 
     def __init__(self):
         super().__init__()
-        self.logger = None
         self.data = None
-        self.config = None
-        self.eval_note = START_EVAL_NOTE
-        self.pertinence = START_EVAL_PERTINENCE
-        self.logger = None
-        self.enabled = True
-        self.is_updating = False
-
         self.short_term_averages = [7, 5, 4, 3, 2, 1]
         self.long_term_averages = [40, 30, 20, 15, 10]
 
-    @classmethod
-    def get_name(cls):
-        return cls.__name__
-
-    def set_logger(self, logger):
-        self.logger = logger
-
     def set_data(self, data):
         self.data = data
-
-    def set_config(self, config):
-        self.config = config
 
     def set_eval_note(self, new_eval_note):
         if self.eval_note + new_eval_note > 1:
@@ -41,32 +24,11 @@ class TAEvaluator:
         else:
             self.eval_note += new_eval_note
 
-    def get_eval_note(self):
-        return self.eval_note
-
-    def get_pertinence(self):
-        return self.pertinence
-
-    def get_is_enabled(self):
-        return self.enabled
-
     def get_is_evaluable(self):
         return not (self.get_is_updating() or self.data is None)
 
-    def get_is_updating(self):
-        return self.is_updating
-
-    def eval(self):
-        self.is_updating = True
-        try:
-            self.eval_impl()
-        except Exception as e:
-            self.logger.error(" Exception in eval_impl(): "+str(e))
-        finally:
-            self.is_updating = False
-
     @abstractmethod
-    def eval_impl(self):
+    def eval_impl(self) -> None:
         raise NotImplementedError("Eval_impl not implemented")
 
     # trend < 0 --> Down trend
