+++ exchanges/simulator/portfolio_trader_simulator.py
@@ -1,0 +1,6 @@
+from exchanges.portfolio_trader import PortfolioTrader
+
+
+class PortfolioTraderSimulator(PortfolioTrader):
+    def __init__(self, config, exchange):
+        super().__init__(config, exchange)
