+++ exchanges/simulator/trader_simulator.py
@@ -1,0 +1,6 @@
+from exchanges.trader import Trader
+
+
+class TraderSimulator(Trader):
+    def __init__(self, config, exchange):
+        super().__init__(config, exchange)
